+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  ### Added
  ### Changed 
  ### Fixed 
+ - Fix the folder audit status event type ([#1193](https://github.com/ehrbase/ehrbase/pull/1193))
 
 ## [0.31.0]
  ### Added