+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  - Removed authorization scopes from endpoints and added support for overwriting controllers ([#1157](https://github.com/ehrbase/ehrbase/pull/1157))
  ### Fixed
  - Fix audit logs location ([#1160](https://github.com/ehrbase/ehrbase/pull/1160))
+ - Fix template id only for composition ([#1186](https://github.com/ehrbase/ehrbase/pull/1186))
 
 ## [0.30.0]
  ### Added