+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
- ### Added
+ ### Added 
  ### Changed 
  ### Fixed 
  - Fix the folder audit status event type ([#1193](https://github.com/ehrbase/ehrbase/pull/1193))
+ - Fix OAuth2 Security Configuration for admin role and management endpoints ([#1196](https://github.com/ehrbase/ehrbase/pull/1196))
+
 
 ## [0.31.0]
  ### Added